cp: add simplify-protocol posthook
